Let all services share the file cache directory (#1183)
* Let all services share the file cache directory

* Disallow destroying/purging a shared cache

* After downloading a file, open the temporary file before moving

Even if someone keeps deleting all cached files, a file can always
be used after downloading it just once. (Previously, this could have
become an infinite loop if the file gets deleted sufficiently quickly
every time.)

* Small changes

+++ cms/db/filecacher.py
@@ -6,6 +6,7 @@
 # Copyright © 2010-2012 Matteo Boscariol <boscarim@hotmail.com>
 # Copyright © 2013 Luca Wehrstedt <luca.wehrstedt@gmail.com>
 # Copyright © 2016 Luca Versari <veluca93@gmail.com>
+# Copyright © 2021 Fabian Gundlach <320pointsguy@gmail.com>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -514,15 +515,13 @@ class FileCacher:
         self._create_directory_or_die(config.temp_dir)
         self._create_directory_or_die(config.cache_dir)
 
-        if service is None:
+        if not self.is_shared():
             self.file_dir = tempfile.mkdtemp(dir=config.temp_dir)
             # Delete this directory on exit since it has a random name and
             # won't be used again.
             atexit.register(lambda: rmtree(self.file_dir))
         else:
-            self.file_dir = os.path.join(
-                config.cache_dir,
-                "fs-cache-%s-%d" % (service.name, service.shard))
+            self.file_dir = os.path.join(config.cache_dir, "fs-cache-shared")
         self._create_directory_or_die(self.file_dir)
 
         # Temp dir must be a subdirectory of file_dir to avoid cross-filesystem
@@ -532,6 +531,10 @@ class FileCacher:
         # Just to make sure it was created.
         self._create_directory_or_die(self.file_dir)
 
+    def is_shared(self):
+        """Return whether the cache directory is shared with other services."""
+        return self.service is not None
+
     @staticmethod
     def _create_directory_or_die(directory):
         """Create directory and ensure it exists, or raise a RuntimeError."""
@@ -540,25 +543,30 @@ class FileCacher:
             logger.error(msg)
             raise RuntimeError(msg)
 
-    def load(self, digest, if_needed=False):
-        """Load the file with the given digest into the cache.
+    def _load(self, digest, cache_only):
+        """Load a file into the cache and open it for reading.
 
-        Ask the backend to provide the file and, if it's available,
-        copy its content into the file-system cache.
+        cache_only (bool): don't open the file for reading.
 
-        digest (unicode): the digest of the file to load.
-        if_needed (bool): only load the file if it is not present in
-            the local cache.
+        return (fileobj): a readable binary file-like object from which
+            to read the contents of the file (None if cache_only is True).
 
-        raise (KeyError): if the backend cannot find the file.
-        raise (TombstoneError): if the digest is the tombstone
+        raise (KeyError): if the file cannot be found.
 
         """
-        if digest == Digest.TOMBSTONE:
-            raise TombstoneError()
         cache_file_path = os.path.join(self.file_dir, digest)
-        if if_needed and os.path.exists(cache_file_path):
-            return
+
+        if cache_only:
+            if os.path.exists(cache_file_path):
+                return
+        else:
+            try:
+                return open(cache_file_path, 'rb')
+            except FileNotFoundError:
+                pass
+
+        logger.debug("File %s not in cache, downloading "
+                     "from database.", digest)
 
         ftmp_handle, temp_file_path = tempfile.mkstemp(dir=self.temp_dir,
                                                        text=False)
@@ -566,9 +574,42 @@ class FileCacher:
                 self.backend.get_file(digest) as fobj:
             copyfileobj(fobj, ftmp, self.CHUNK_SIZE)
 
+        if not cache_only:
+            # We allow anyone to delete files from the cache directory
+            # self.file_dir at any time. Hence, cache_file_path might no
+            # longer exist an instant after we create it. Opening the
+            # temporary file before renaming it circumvents this issue.
+            # (Note that the temporary file may not be manually deleted!)
+            fd = open(temp_file_path, 'rb')
+
         # Then move it to its real location (this operation is atomic
         # by POSIX requirement)
         os.rename(temp_file_path, cache_file_path)
+
+        logger.debug("File %s downloaded.", digest)
+
+        if not cache_only:
+            return fd
+
+    def cache_file(self, digest):
+        """Load a file into the cache.
+
+        Ask the backend to provide the file and store it in the cache for the
+        benefit of future accesses, unless the file is already cached.
+        Note that the cached file might still be deleted at any time, so it
+        cannot be assumed to actually exist after this function completes.
+        Always use the get_file* functions to access a file.
+
+        digest (unicode): the digest of the file to get.
+
+        raise (KeyError): if the file cannot be found.
+        raise (TombstoneError): if the digest is the tombstone
+
+        """
+        if digest == Digest.TOMBSTONE:
+            raise TombstoneError()
+
+        self._load(digest, True)
 
     def get_file(self, digest):
         """Retrieve a file from the storage.
@@ -592,19 +633,10 @@ class FileCacher:
         """
         if digest == Digest.TOMBSTONE:
             raise TombstoneError()
-        cache_file_path = os.path.join(self.file_dir, digest)
 
         logger.debug("Getting file %s.", digest)
 
-        if not os.path.exists(cache_file_path):
-            logger.debug("File %s not in cache, downloading "
-                         "from database.", digest)
-
-            self.load(digest)
-
-            logger.debug("File %s downloaded.", digest)
-
-        return open(cache_file_path, 'rb')
+        return self._load(digest, False)
 
     def get_file_content(self, digest):
         """Retrieve a file from the storage.
@@ -663,33 +695,6 @@ class FileCacher:
             with open(dst_path, 'wb') as dst:
                 copyfileobj(src, dst, self.CHUNK_SIZE)
 
-    def save(self, digest, desc=""):
-        """Save the file with the given digest into the backend.
-
-        Use to local copy, available in the file-system cache, to store
-        the file in the backend, if it's not already there.
-
-        digest (unicode): the digest of the file to load.
-        desc (unicode): the (optional) description to associate to the
-            file.
-
-        raise (TombstoneError): if the digest is the tombstone
-
-        """
-        if digest == Digest.TOMBSTONE:
-            raise TombstoneError()
-        cache_file_path = os.path.join(self.file_dir, digest)
-
-        fobj = self.backend.create_file(digest)
-
-        if fobj is None:
-            return
-
-        with open(cache_file_path, 'rb') as src:
-            copyfileobj(src, fobj, self.CHUNK_SIZE)
-
-        self.backend.commit_file(fobj, digest, desc)
-
     def put_file_from_fobj(self, src, desc=""):
         """Store a file in the storage.
 
@@ -738,16 +743,20 @@ class FileCacher:
 
             cache_file_path = os.path.join(self.file_dir, digest)
 
-            if not os.path.exists(cache_file_path):
-                os.rename(dst.name, cache_file_path)
-            else:
-                os.unlink(dst.name)
+            # Store the file in the backend. We do that even if the file
+            # was already in the cache
+            # because there's a (small) chance that the file got removed
+            # from the backend but somehow remained in the cache.
+            # We read from the temporary file before moving it to
+            # cache_file_path because the latter might be deleted before
+            # we get a chance to open it.
+            with open(dst.name, 'rb') as src:
+                fobj = self.backend.create_file(digest)
+                if fobj is not None:
+                    copyfileobj(src, fobj, self.CHUNK_SIZE)
+                    self.backend.commit_file(fobj, digest, desc)
 
-        # Store the file in the backend. We do that even if the file
-        # was already in the cache (that is, we ignore the check above)
-        # because there's a (small) chance that the file got removed
-        # from the backend but somehow remained in the cache.
-        self.save(digest, desc)
+            os.rename(dst.name, cache_file_path)
 
         return digest
 
@@ -843,6 +852,8 @@ class FileCacher:
     def purge_cache(self):
         """Empty the local cache.
 
+        This function must not be called if the cache directory is shared.
+
         """
         self.destroy_cache()
         if not mkdir(config.cache_dir) or not mkdir(self.file_dir):
@@ -855,7 +866,11 @@ class FileCacher:
         Nothing that could have been created by this object will be
         left on disk. After that, this instance isn't usable anymore.
 
+        This function must not be called if the cache directory is shared.
+
         """
+        if self.is_shared():
+            raise Exception("You may not destroy a shared cache.")
         rmtree(self.file_dir)
 
     def list(self):
